add more logging facility for idunn
The main thing is that we are now able to output json logs (for easier
ELK integration).

To enable it, the `IDUNN_LOG_JSON` must be set to 1.

`LOG_FORMAT` can be used to defined the logged fields / format
`LOG_LEVEL_BY_MODULE` can be set to define the log level for each module

Gunicorn now also output its log as json (and this is not configurable,
it's set in the Dockerfile)